+++ ReleaseNotes.md
@@ -1,3 +1,8 @@
+# Release 1.1.1
+- Add Javadoc and Groovydoc to Build
+- Correct Pom generation for uploadArchives
+- Move Signing-Settings from init.gradle to PublishPlugin
+
 # Release 1.1.0
 - New Funktionality for automatic jar and pom signing
 - Add Documentation for gitHub Release